Refactor and enhance game systems and GUI integration
Refactored `init()` in `main.cpp` to reorder initialization and integrate `Gui` with `InputManager`. Transitioned `Inventory` resource handling from `int` to `float` for fractional values. Updated `Structure` and `SolarPanel` to support time-based resource management using `dt`.

Integrated `ImGui` for GUI rendering, adding game speed control, resource visualization, and build menus. Enhanced `InputManager` to handle tool-based interactions and ensure compatibility with `ImGui`.

Improved `Renderer` to differentiate selection highlights. Updated `Map` to handle invalid structure placement gracefully. Enhanced `World` to calculate total energy and capacity dynamically. Introduced time-based updates across systems for smoother gameplay.

+++ src/entities/Inventory.h
@@ -4,19 +4,19 @@
 
 class Inventory {
 private:
-    std::unordered_map<MaterialType, int> items;
-    std::unordered_map<MaterialType, int> maxCapacities;
+    std::unordered_map<MaterialType, float> items;
+    std::unordered_map<MaterialType, float> maxCapacities;
 
 public:
     void init();
 
-    void setMaxCapacity(MaterialType type, int capacity);
-    int getMaxCapacity(MaterialType type) const;
+    void setMaxCapacity(MaterialType type, float capacity);
+    float getMaxCapacity(MaterialType type) const;
 
-    bool addResource(MaterialType type, int amount); 
-    bool subResource(MaterialType type, int amount);     
-    bool hasResource(MaterialType type, int amount) const;
-    int getAmount(MaterialType type) const;
+    bool addResource(MaterialType type, float amount); 
+    bool subResource(MaterialType type, float amount);     
+    bool hasResource(MaterialType type, float amount) const;
+    float getAmount(MaterialType type) const;
 
     void clear();
 };
